2- copies strings

+++ 0x06-pointers_arrays_strings/2-strncpy.c
@@ -10,7 +10,7 @@
 
 char *_strncpy(char *dest, char *src, int n)
 {
-		size_t i, len;
+		int i, len;
 
 		len = strlen(dest);
 		for (i = 0 ; i < n && src[i] != '\0'; i++)
